Migrate properties
