--- web/html/src/manager/content-management/shared/components/panels/properties/properties-form.js
@@ -1,53 +1,0 @@
-// @flow
-import * as React from 'react';
-import {Form} from "components/input/Form";
-import {Text} from "components/input/Text";
-
-import type {ProjectPropertiesType} from '../../../type/project.type.js';
-
-type Props = {
-  properties: ProjectPropertiesType,
-  errors: Object,
-  onChange: Function,
-  editing?: boolean
-};
-
-const PropertiesForm = (props: Props) => (
-  <Form
-    model={props.properties}
-    errors={props.errors}
-    onChange={model => {
-      props.onChange(model);
-    }}
-  >
-    <div className="row">
-      <Text
-        required
-        name="name"
-        label={t("Name")}
-        labelClass="col-md-2"
-        divClass="col-md-10"
-      />
-    </div>
-    <div className="row">
-      <Text
-        required
-        name="label"
-        label={t("Label")}
-        labelClass="col-md-2"
-        divClass="col-md-10"
-        disabled={props.editing}
-      />
-    </div>
-    <div className="row">
-      <Text
-        name="description"
-        label={t("Description")}
-        labelClass="col-md-2"
-        divClass="col-md-10"
-      />
-    </div>
-  </Form>
-);
-
-export default PropertiesForm;
